+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Main (unreleased)
 
+- [ENHANCEMENT] Traces: Improved pod association in PromSD processor (@mapno)
+
 # v0.21.1 (2021-11-18)
 
 - [BUGFIX] Fix panic when using postgres_exporter integration (@saputradharma)